Consolidate header and footer bars into SelectedObjectPanel
